+++ packages/react/test/specs/commonTests/implementsCollectionShorthandProp.tsx
