+++ packages/react/test/specs/commonTests/implementsWrapperProp.tsx
