+++ packages/react/test/specs/commonTests/isConformant.tsx
