+++ packages/react/test/specs/commonTests/isExportedAtTopLevel.tsx
