+++ packages/react/test/specs/components/Avatar/Avatar-test.tsx
