+++ packages/react/test/specs/components/Chat/ChatItem-test.tsx
@@ -1,4 +1,4 @@
-import { isConformant, implementsShorthandProp } from 'test/specs/commonTests'
+import { isConformant, implementsShorthandProp } from '../../commonTests'
 import ChatItem from 'src/components/Chat/ChatItem'
 import Box from 'src/components/Box/Box'
