+++ packages/react/test/specs/components/Chat/ChatMessage-test.tsx
@@ -1,7 +1,7 @@
-import { handlesAccessibility, implementsShorthandProp, isConformant } from 'test/specs/commonTests'
+import { handlesAccessibility, implementsShorthandProp, isConformant } from '../../commonTests'
 
 import ChatMessage from 'src/components/Chat/ChatMessage'
-import { chatMessageBehavior } from 'src/lib/accessibility'
+import { chatMessageBehavior } from '../../../../src/lib/accessibility'
 import { AccessibilityDefinition } from 'src/lib/accessibility/types'
 import Text from 'src/components/Text/Text'
 import Box from 'src/components/Box/Box'
